The DevEvalFileProduct class

Package: R.devices
Class DevEvalFileProduct

character
~~|
~~+--BasicObject
~~~~~~~|
~~~~~~~+--DevEvalProduct
~~~~~~~~~~~~|
~~~~~~~~~~~~+--DevEvalFileProduct

Directly known subclasses:

public class DevEvalFileProduct
extends DevEvalProduct

A DevEvalFileProduct is a DevEvalProduct referring to a image file created by devEval().

DevEvalFileProduct(filename=NULL, path=NULL, ...)

Arguments

filename, path

The filename and the optional path of the image file product.

...

Additional arguments passed to DevEvalProduct.

Fields

The following (virtual; calculate on-the-fly) fields are available:

pathname: the (relative) pathname of the image file, e.g. 'figures/foo,a,b.png'. This can be used to link to image files in for instance HTML and Markdown documents.

path, the (relative) path to the image file, e.g. 'figures/'

filename: the filename ("basename") of the image file, e.g. 'foo,a,b.png'

fullname: the fullname (the filename without the filename extension), e.g. 'foo,a,b'. It is recommended to use this when including images in LaTeX documents that use the \usepackage{graphicx} package.

name: the part of the fullname before the first comma, e.g. 'foo'

tags: the part of the fullname after the first comma, e.g. 'a,b'

dataURI: the Base64-encoded Data URI representation of the image, e.g. 'data:image/png;base64,iVBORw0KGgoAAAA...'. This can be used to include ("inlining") image files in for instance self-contained HTML and Markdown documents.

data: the character content of the image file. This can be used to include ("inlining") WebGL HTML-based image files in for instance self-contained HTML and Markdown documents.

See also

Author

Henrik Bengtsson