Package R.devices

Functions for creating plots and image files in a unified way regardless of output format (EPS, PDF, PNG, SVG, TIFF, WMF, etc.). Default device options as well as scales and aspect ratios are controlled in a uniform way across all device types. Switching output format requires minimal changes in code. This package is ideal for large-scale batch processing, because it will never leave open graphics devices or incomplete image files behind, even on errors or user interrupts.

To get started

Vignette 'R.devices overview'

toEPS(), toPDF(), toPNG(), ... - evaluate graphics code and atomically save plot to a figure file.

How to cite this package

To cite this package, please use:

@Manual{,
  title = {R.devices: Unified Handling of Graphics Devices},
  author = {Henrik Bengtsson},
  year = {2024},
  note = {R package version 2.17.2,
https://github.com/HenrikBengtsson/R.devices},
  url = {https://henrikbengtsson.github.io/R.devices/},
}

Author

Henrik Bengtsson [aut, cre, cph].

License

LGPL (>= 2.1).