EPS graphics device

Device driver for Encapsulated Postscript. This driver is the same as the postscript driver where some arguments have different default values.

eps(file="Rplot%03d.eps", width=7, height=7, horizontal=FALSE, paper="special",
  onefile=FALSE, ...)

Arguments

file

Default file name (pattern).

width, height

The width and height of the figure.

horizontal

If FALSE, an horizontal EPS file is created, otherwise a portrait file is created.

paper

A character string specifying the paper type. Overrides the default of postscript().

onefile

Not used.

...

Other arguments accepted by postscript().

Value

A plot device is opened; nothing is returned.

See also

Author

Henrik Bengtsson

Examples

if (FALSE) {
  eps("foo.eps", width=7, height=7)

  # is identical to

  postscript("foo.eps", width=7, height=7, onefile=TRUE, horizontal=FALSE)

  # and

  dev.print(eps, "foo.eps", ...)

  # is identical to

  dev.print(postscript, "foo.eps", onefile=TRUE, horizontal=FALSE, paper="special", ...)
}