Methods for creating image files of a specific format

Methods for creating image files of a specific format.

toBMP(name, ...)
   toPDF(name, ...)
   toPNG(name, ...)
   toSVG(name, ...)
   toTIFF(name, ...)
   toEMF(name, ..., ext="emf")
   toWMF(name, ..., ext="wmf")

   toFavicon(..., name="favicon", ext="png",
             field=getDevOption("favicon", "field", default="htmlscript"))

   <!-- % Interactive/screen devices -->
   toDefault(name, ...)
   toQuartz(name, ...)
   toX11(name, ...)
   toWindows(name, ...)

   <!-- % Interactive/screen devices from other packages -->
   toCairoWin(name, ...)
   toCairoX11(name, ...)
   <!-- %% toJavaGD(name, ...) -->

   <!-- % Interactive/screen device for RStudio -->
   toRStudioGD(name, ..., .allowUnknownArgs = TRUE)

Arguments

name

A character string specifying the name of the image file.

..., .allowUnknownArgs

Additional arguments passed to devEval(), e.g. tags and aspectRatio.

ext, field

Passed to devEval().

Value

Returns by default the DevEvalProduct. For toFavicon() the default return value is a character string.

Windows Metafile Format

Both toEMF() and toWMF() use the exact same graphics device (win.metafile()) and settings. They only differ by filename extension. The win.metafile() device function exists on Windows only; see the grDevices package for more details.

Author

Henrik Bengtsson

See also